requirements.txt
langchain>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
langchain-community>=0.3.0,<1.0.0
langchain-core>=1.0.0,<2.0.0
langsmith>=0.3.0,<1.0.0
requests==2.32.5
pandas==2.2.3
matplotlib==3.10.1
PyYAML>=6.0.1
flask==3.1.2
flask-cors==6.0.1
docker>=6.1.3
scipy>=1.11.4
biopython==1.86  # PubMed检索依赖
python-dotenv==1.2.1
fastmcp==3.0.2
chromadb>=0.5.5
langchain-chroma>=0.1.0